src/requirements.txt
flake8==6.0.0
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==0.21.0
